+++ docs/docs/release-notes.md
@@ -6,6 +6,17 @@ sidebar_position: 200
 
 ## Unreleased
 
+## features
+- **[API]**: We are working on a v2 version of the API, we have already migrated the endpoints:
+  - `/api/data_product_lifecycles` -> `/api/v2/configuration/data_product_lifecycles`
+  - `/api/data_product_settings` -> `/api/v2/configuration/data_product_settings`
+  - `/api/data_product_types` -> `/api/v2/configuration/data_product_types`
+  - `/api/domains` -> `/api/v2/configuration/domains`
+  - `/api/environments` -> `/api/v2/configuration/environments`
+  - `/api/platforms` -> `/api/v2/configuration/platforms`
+  - `/api/tags` -> `/api/v2/configuration/tags`
+  - `/api/theme_settings` -> `/api/v2/configuration/theme_settings`
+
 ## 0.4.1
 
 ### features